job tag command split into apply and unset
