+++ command/job_tag.go
@@ -4,216 +4,34 @@
 package command
 
 import (
-	"fmt"
 	"strings"
 
-	"github.com/posener/complete"
+	"github.com/mitchellh/cli"
 )
 
 type JobTagCommand struct {
 	Meta
 }
 
-func (c *JobTagCommand) Help() string {
-	helpText := `
-Usage: nomad job tag apply [options] <jobname>
+func (c *JobTagCommand) Name() string { return "job tag" }
 
-  Save a job version to prevent it from being garbage-collected and allow it to
-  be diffed and reverted by name.
-  
-  Example usage:
- 
-    nomad job tag apply -name "My Golden Version" \
-		-description "The version we can roll back to if needed" <jobname>
-
-    nomad job tag apply -version 3 -name "My Golden Version" <jobname>
-
-  The first of the above will tag the latest version of the job, while the second
-  will specifically tag version 3 of the job.
-
-Tag Specific Options:
-
-  -name <version-name>
-    Specifies the name of the version to tag. This is a required field.
-
-  -description <description>
-    Specifies a description for the version. This is an optional field.
-
-  -version <version>
-    Specifies the version of the job to tag. If not provided, the latest version
-    of the job will be tagged.
-
-
-General Options:
-
-  ` + generalOptionsUsage(usageOptsNoNamespace) + `
-`
-	return strings.TrimSpace(helpText)
+func (c *JobTagCommand) Run(args []string) int {
+	return cli.RunResultHelp
 }
 
 func (c *JobTagCommand) Synopsis() string {
-	return "Save a job version to prevent it from being garbage-collected and allow it to be diffed and reverted by name."
+	return "Manage job version tags"
 }
 
-func (c *JobTagCommand) AutocompleteFlags() complete.Flags {
-	return mergeAutocompleteFlags(c.Meta.AutocompleteFlags(FlagSetClient),
-		complete.Flags{
-			"-name":        complete.PredictAnything,
-			"-description": complete.PredictAnything,
-			"-version":     complete.PredictNothing,
-		})
-}
-
-func (c *JobTagCommand) AutocompleteArgs() complete.Predictor {
-	return complete.PredictNothing
-}
-
-func (c *JobTagCommand) Name() string { return "job tag apply" }
-
-func (c *JobTagCommand) Run(args []string) int {
-	var name, description, versionStr string
-
-	flags := c.Meta.FlagSet(c.Name(), FlagSetClient)
-	flags.Usage = func() { c.Ui.Output(c.Help()) }
-	flags.StringVar(&name, "name", "", "")
-	flags.StringVar(&description, "description", "", "")
-	flags.StringVar(&versionStr, "version", "", "")
-
-	if err := flags.Parse(args); err != nil {
-		return 1
-	}
-
-	if len(flags.Args()) != 1 {
-		c.Ui.Error("This command takes one argument: <job>")
-		c.Ui.Error(commandErrorText(c))
-		return 1
-	}
-
-	var job = flags.Args()[0]
-
-	if job == "" {
-		c.Ui.Error("A job name is required")
-		c.Ui.Error(commandErrorText(c))
-		return 1
-	}
-
-	if name == "" {
-		c.Ui.Error("A version tag name is required")
-		c.Ui.Error(commandErrorText(c))
-		return 1
-	}
-
-	// Get the HTTP client
-	client, err := c.Meta.Client()
-	if err != nil {
-		c.Ui.Error(fmt.Sprintf("Error initializing client: %s", err))
-		return 1
-	}
-
-	// Check if the job exists
-	jobIDPrefix := strings.TrimSpace(job)
-	jobID, _, err := c.JobIDByPrefix(client, jobIDPrefix, nil)
-	if err != nil {
-		c.Ui.Error(err.Error())
-		return 1
-	}
-
-	_, err = client.Jobs().TagVersion(jobID, versionStr, name, description, nil)
-	if err != nil {
-		c.Ui.Error(fmt.Sprintf("Error tagging job version: %s", err))
-		return 1
-	}
-
-	return 0
-}
-
-// #region unset
-
-type JobTagUnsetCommand struct {
-	Meta
-}
-
-func (c *JobTagUnsetCommand) Help() string {
+func (c *JobTagCommand) Help() string {
 	helpText := `
-  TODO: Write help text
-  `
+Usage: nomad job tag <subcommand> [options] [args]
+
+  This command is used to manage tags for job versions. It has subcommands
+  for applying and unsetting tags.
+
+For more information on a specific subcommand, run:
+  nomad job tag <subcommand> -h
+`
 	return strings.TrimSpace(helpText)
 }
-
-func (c *JobTagUnsetCommand) Synopsis() string {
-	return "Remove a tag from a job version."
-}
-
-func (c *JobTagUnsetCommand) AutocompleteFlags() complete.Flags {
-	return mergeAutocompleteFlags(c.Meta.AutocompleteFlags(FlagSetClient),
-		complete.Flags{
-			// "-version": complete.PredictNothing,
-		})
-}
-
-func (c *JobTagUnsetCommand) AutocompleteArgs() complete.Predictor {
-	return complete.PredictNothing
-}
-
-func (c *JobTagUnsetCommand) Name() string { return "job tag unset" }
-
-func (c *JobTagUnsetCommand) Run(args []string) int {
-	var name string
-
-	flags := c.Meta.FlagSet(c.Name(), FlagSetClient)
-	flags.Usage = func() { c.Ui.Output(c.Help()) }
-	flags.StringVar(&name, "name", "", "")
-
-	if err := flags.Parse(args); err != nil {
-		return 1
-	}
-
-	if len(flags.Args()) != 1 {
-		c.Ui.Error("This command takes one argument: <job>")
-		c.Ui.Error(commandErrorText(c))
-		return 1
-	}
-
-	var job = flags.Args()[0]
-
-	if job == "" {
-		c.Ui.Error(
-			"A job name is required",
-		)
-		c.Ui.Error(commandErrorText(c))
-		return 1
-	}
-
-	if name == "" {
-		c.Ui.Error(
-			"A version tag name is required",
-		)
-		c.Ui.Error(commandErrorText(c))
-		return 1
-	}
-
-	// Get the HTTP client
-	client, err := c.Meta.Client()
-	if err != nil {
-		c.Ui.Error(fmt.Sprintf("Error initializing client: %s", err))
-		return 1
-	}
-
-	// Check if the job exists
-	jobIDPrefix := strings.TrimSpace(job)
-	jobID, _, err := c.JobIDByPrefix(client, jobIDPrefix, nil)
-	if err != nil {
-		c.Ui.Error(err.Error())
-		return 1
-	}
-
-	_, err = client.Jobs().UntagVersion(jobID, name, nil)
-	if err != nil {
-		c.Ui.Error(fmt.Sprintf("Error tagging job version: %s", err))
-		return 1
-	}
-
-	return 0
-}
-
-// #endregion unset
